sidecar/requirements.txt
fastapi>=0.116.0,<1.0.0
httpx>=0.28.0,<1.0.0
openai>=2.30.0,<3.0.0
openai-agents>=0.13.4,<1.0.0
openrouter>=0.8.1,<1.0.0
python-dotenv>=1.2.0,<2.0.0
python-multipart>=0.0.20,<1.0.0
sqlalchemy>=2.0.0,<3.0.0
uvicorn>=0.35.0,<1.0.0
